google: add back re-run
I did these re-runs because the settings on the kubernetes
event exporting was dropping some events, and I do not
think that is appropriate data to use for a publication.
I am still going to run one more final study on gKE
that only tests the regular pulls using their registry.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>